Return categorical values which differ between data and their data dictionary

Generates a tibble that reports any categorical variable which categorical possible observation (the combination of 'variable' and 'name' in 'Categories' sheet) is actually observed in the data observations. Likewise, the actual observations are evaluated to report any categorical variable which occurrence is not in the data dictionary categorical variables. This tibble can be used to assist the user in the assessment of the data structure, fields investigation (mandatory or not), coherence across elements and taxonomy or standard evaluation.

check_dataset_categories(data, data_dict = NULL)

Arguments

data

A tibble identifying the input data observations.

data_dict

A list of tibble(s) representing meta data to be evaluated.

Value

A tibble providing categorical values which differ between data and their data dictionary.

Details

A data dictionary-like structure must be a list of at least one or two data frame or data frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research stardards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and can be associated to a data dictionary. If not, a minimum workable data dictionary can always be generated, when any column will be reported, and any factor column will be analysed as categorical variable (the column 'levels' will be created for that. In addition, the dataset may follow Maelstrom research stardards, and its content can be evaluated accordingly, such as naming convention restriction, or id columns declaration (which full completeness is mandatory.

Examples

if (FALSE) {
# Example 1: yyy yyy yyy.
}