Generate a report as an Excel spreadsheet of a study-specific data dictionary

Generates an Excel spreadsheet report for a data dictionary for each variable to facilitate the assessment of input data. This report can be used to assist the user in the assessment of the data structure, fields investigation (mandatory or not), coherence across elements and taxonomy, or standard evaluation. The summary associated provides data dictionary composition, and evaluates coherence between elements.

data_dict_evaluate(data_dict, taxonomy = NULL, as_mlstr_data_dict = TRUE)

Arguments

data_dict

A list of tibble(s) representing meta data to be evaluated.

taxonomy

A tibble identifying the scheme used for variables classification as a tibble.

as_mlstr_data_dict

Whether the output data dictionary has a simple data dictionary structure or not (meaning has a Maelstrom data dictionary structure, compatible with Maelstrom ecosystem such as Opal environment). FALSE by default.

Value

A list of tibbles of report for one study-specific data dictionary.

Details

A data dictionary-like structure must be a list of at least one or two data frame or data frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research stardards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A taxonomy must be a data frame or data frame extension (e.g. a tibble). The taxonomy must be compatible with (and generally extracted from) an Opal environment, and must contain at least 'taxonomy', 'vocabulary' and 'terms' to work with some specific functions. In addition, the taxonomy may follow Maelstrom research stardards, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

Examples