Generate a web application (bookdown) report of list of study-specific datasets

Generates a visual report for a study-specific dataset in an HTML bookdown document, showing descriptive statistics for each study-specific variable to facilitate the assessment of input data. Statistics and figures are generated according to their valueTypes. This report can be used to assist the user in the assessment of the data structure, fields investigation (mandatory or not), coherence across elements and taxonomy or standard evaluation. The summaries and figures associated provide dataset composition, with observation repartition and descriptive statistics.

dataset_visualize(
  dataset = NULL,
  data_dict = NULL,
  group_by = NULL,
  to,
  out = "ggplot2",
  .keep_files = TRUE
)

Arguments

dataset

A tibble identifying the input data observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset. Automatically generated if not provided.

group_by

A character string of one column in the dataset that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

to

A character string identifying the folder path where the bookdown report will be saved.

out

parameter that specifies the graphical outputs expected in the report. can be 'ggplot2'.

.keep_files

whether to keep the R-markdown files. FALSE by default.

Value

A bookdown folder containing files in the specified output folder. To open the file in browser, open 'index.html'. Or use open_visual_report()

Details

A study must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being datasets. The name of each tibble will be use as the reference name of the dataset. A data dictionary-like structure must be a list of at least one or two data frame or data frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research stardards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A dataset must be a data frame or data frame extension (e.g. a tibble) and can be associated to a data dictionary. If not, a minimum workable data dictionary can always be generated, when any column will be reported, and any factor column will be analysed as categorical variable (the column 'levels' will be created for that. In addition, the dataset may follow Maelstrom research stardards, and its content can be evaluated accordingly, such as naming convention restriction, or id columns declaration (which full completeness is mandatory.

See also

Examples