Upload files to an Opal environment

Uploads files from local to Opal environment. It is a wrapper of opalr::opal.file_upload().

opal_files_push(opal, from, to)

Arguments

opal

Opal login attributes.

from

A character string of a path where the files will be taken from in R.

to

A character string of a path where the files will be placed to in Opal.

Value

Folder(s) containing files coming from the user R environment in Opal. The path to Opal needs to be pasted with Opal absolute path.

Details

The user must be allowed to interact with their Opal. The errors provided may be associated to the handler or the read/write permissions to Opal. The user must have adequate credentials to interact with their Opal environment.

See also

Examples